src/requirements.txt
Eve==0.8
Flask-PyMongo==0.5.2
passlib==1.7.1
PyJWT==1.6.4
python-dotenv==0.9.1
